rename _py_op_cli.py to _op_cli.py
